+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Reversing a blend in progress respects asymmetric blend times.
 - Regression fix: 3rdPersonFollow did not keep keep player in view when Z damping is high
 - Regression fix: CmPostProcessing and CmVolumeSettings components setting Depth of Field did not work correctly with Framing Transposer. 
+- Bugfix: CinemachineCollider's displacement damping was being calculated in world space instead of camera space.
 
 
 ## [2.7.3] - 2021-03-24